Removing random NPC despawn (#428)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
NPCs are already limited to respawning only up to 130, and though this
supposedly optimizes their routes, it is kind of annoying as they tend
to randomly disappear. From my tests of 10+ minutes nothing went wrong,
although if you follow one around you'll notice he's walking in circles
- but that doesn't matter, it's better than them popping out of
existence.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
It will make it easier to handle and deal with NPCs, kidnap them, and
such.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog
:cl:
tweak: tweaked NPCs
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->