[MAJOR MAPPING] Pentex Headquarters - No more warehouse! (#667)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This pull request removes the old Pentex warehouse and replaces it with
an entirely new, built from scratch Pentex Facility (most likely owned
by Endron). Here, there are many rooms, ranging from a laboratory, to
cells, to the executive's office with a hidden secret behind the
blast-door.

## Why It's Good For The Game

- This can add a major RP hub for BSD to properly roleplay with others
outside of just messing with the Glasswalkers or Kindred.
- Pentex now has somewhat of a presence on the map and serves as the
framework for when someone wants to add the roles.
- The BSD totem is now properly secure and requires a bit of effort to
get to.

## Changelog

del: Entirely removes the old Pentex Warehouse
add: Adds a new facility, accessable by area portals (stairs, as it is
on a hill).
add: New areas for Pentex affiliated individuals, mainly BSD to RP in
with themselves and others.


![image](https://github.com/user-attachments/assets/6df8540d-1c52-4005-84c5-720120dfeaf0)

![image](https://github.com/user-attachments/assets/4cf596cd-1bd7-4e75-9cfb-4e54841ddecb)

![image](https://github.com/user-attachments/assets/e16b280c-66dc-4c81-9759-8a6e6ff7d53c)

![image](https://github.com/user-attachments/assets/a2992b13-e252-4b62-b6a5-d351ebd08356)

![image](https://github.com/user-attachments/assets/0648b573-b050-4483-9649-194e40afe666)

![image](https://github.com/user-attachments/assets/ea7305e2-4b51-4a15-a0d7-3583507bb770)