Mytherceria disorient runtime fixes (#620)
## About The Pull Request
These two abilities are supposed to disorient another persons hud,
either from casting myth 4 on them or from being caught in a spin trap,
however if they don't have a client, like NPCs or simple_mobs, they
don't have a hud and this then throws a runtime when they are caught in
these. Adds checks for both to make sure that the code only triggers
when someone with a client (and therefor, a hud) is targetted.


![bild](https://github.com/user-attachments/assets/f23f088d-03e0-49ae-bbbf-424fd7751b29)
## Why It's Good For The Game
Runtimes are stinky and complicate things

## Changelog

Yet another runtime cleanup, nothing of note for players to know or care
about.